install.sh: add some checks

+++ install.sh
@@ -1,10 +1,19 @@
 #!/bin/sh
 
+if [ -d ~/.vim ]; then
+  echo '~/.vim already exists, exiting.'
+  exit 1
+fi
+
 echo 'Cloning vim'
 git clone git://github.com/gwarf/vim.git ~/.vim
 
-echo 'Symlinking vimrc'
-ln -s ~/.vim/vimrc ~/.vimrc
+if [ -e ~/.vimrc ]; then
+  echo '~/.vimrc already; skip linking to new one'
+else
+  echo 'Symlinking vimrc'
+  ln -s ~/.vim/vimrc ~/.vimrc
+  echo 'Installing plugins'
+  vim +PlugInstall
+fi
 
-echo 'Installing plugins'
-vim +PlugInstall
